[TASK] Removed locallang class variables
The class variables containing the language labels were replaced with
direct calls to the pi_getLL() method.

This reduces the amount of unused calls to pi_getLL().

Additionally the code that checks if labels for no vote or one vote are
present was restructured and a possible empty label prevented.

Finally the unused translation for the submit_js_linktext key war removed
from the locallang.xml file.